app/requirements.txt
# api
googlemaps==3.0.2
responses==0.10.2
requests==2.20.0

# data manipulation
numpy==1.15.3
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.7

# databases
psycopg2==2.7.6.1
psycopg2-binary-2.7.6.1

# parsing
bs4==0.0.1

pdfkit-0.6.1
tqdm-4.28.1